+++ packages/swarm-coordination-registry/src/swarm/registry.rs
@@ -403,7 +403,7 @@ impl Registry {
             let stats = swarm.metadata();
 
             metrics.total_complete += u64::from(stats.complete);
-            metrics.total_downloaded += u64::from(stats.downloaded);
+            metrics.total_downloaded += stats.downloaded;
             metrics.total_incomplete += u64::from(stats.incomplete);
             metrics.total_torrents += 1;
         }
