+++ packages/torrent-repository-benchmarking/src/repository/rw_lock_std_mutex_tokio.rs
@@ -93,7 +93,7 @@ where
         for entry in entries {
             let stats = entry.lock().await.get_swarm_metadata();
             metrics.total_complete += u64::from(stats.complete);
-            metrics.total_downloaded += u64::from(stats.downloaded);
+            metrics.total_downloaded += stats.downloaded;
             metrics.total_incomplete += u64::from(stats.incomplete);
             metrics.total_torrents += 1;
         }
